remove the StrCpy, StrCat, StrLen functions
they do the same as the ones we get from the compiler
except the one with the compiler are crazily optimized usually so they
should be used instead of these.
Also add a workaround for the double free for the horn and warcry
sounds.
This sound system is totally broken and should be redone.