2nd round of Str functions removal
and a better workaround for the free of horn & warcry
the pointers are not tested directly anymore, but I got a serious
problem with the Mix_Chunk struct which can't be modified, if you add an
int keep at its end, it creates havoc because it doesn't seem to be seen
the same way from different points in the code, which makes no sense at
all! I'll need more tests for that, but for now this workaround is still
better.