finally some proper fixes!
2 fixes in one time, both were linked. Finally it's stdafx.h which
changes the size of stuct Mix_Chunk, 32 if not included, 21 if included!
The danger is that it triggers no warning anywhere, and if passing a
Mix_Chunk from a 21 part to a 32 part you are sure there will be
problems!
One such problem was the weird sdlChannels which could only be a
pointer, it was because of this! The other problem was trying to add an
int keep at the end of Mix_Chunk which created all kinds of craziness.
I didn't pinpoint the exact line responsible, it's probably somewhere
inside CSBTypes.h, but for now it's enough to know you need to include
stdafx.h where types are shared.