a more modern makefile
with automatic dependancies generation, silent compiling and linking
much shorter and more precise for dependancies.
Notice that for now I'll stay way from cmake because there is always an
environment variable which is supposed to be here but which is not for
some obscure reason, and in the end you loose time using cmake when it's
supposed to make things easier. At least with a makefile you can run
anything and the way you like !