add sdl_sound
The handling of external sounds to play is better because of sdl_sound,
we can accept mp3 directly now.
The audio system is quite different, previously the mixer was opened in
mono, 5120 Hz, 8 bits and it supported only that, now it opens the mixer
with its preferred audio parameters, only forcing the format to 16 bits
integers to make sound mixing easier.
Sound is still mono for now, didn't try any panning yet.
Also the makefile now creates an obj directory.