get rid of pragma pack(1) !
My 1st reflex was there was a very good reason for this thing, so I
kept it. But after testing, I coudln't find it! The game correctly loads
its saves even without it! In this case, get rid of it, too dangerous,
and the code will work better without it!