Monster walking sounds + panning
and attenuation, to place sounds in space
The walls are ignored for sound propagation, but attenuation is
proportional to the square of the distance, so you don't hear very far
anyway, and it can be fun to hear things just on the other side of a
wall!
There are still monsters without sounds because of the naming of the
constants for monsters identification, some are just impossible to find.
I'll need to get there in game, but it might take a little time.

Also known bug: the same custom sound can't be played twice at the same
time, example : 2 warcries very close to each other. The reason is
because the data comes from the same Sound_Sample, so it can't be shared
between the 2 sources without problems ! I'll have to think of something
for that.