add a runtime icon
it's mainly for the linux version for now since the windows version has
an icon merged into the exe, and the linux version can't be launched
outside its data directory for now.
The icon is in bmp instead of png so that I don't need to use SDL_image
for now, but I might switch to SDL_image later maybe. The bmp file
remains small anyway, 33 kb. It's not in git, but it's in the archives
with the data files and the sounds.