Finally add a command to load a savegame from the menu!
It's still primitive, there are errors appearing when hovering over some
parts of the file selector, but it's usable, and it works. Also when loading a savegame
you are returned to the initial screen of the game, you should click the
red button to load a game, but you won't be asked which game this time.