add a proper way to detect "dm rules"
and add a menu entry for it just in case.
The official dm dungeon.dat should be detected all the time now, tested
with pc version 3.4, and fm-town version. Added a menu entry with "dm
rules" just in case. If dm rules are disabled, you take a massive
penalty when reincarnating!