weird graphics.dat handling...
avoid to loop on an error message box when trying to load the pc version
graphics.dat. It still crashes in the end, apparently csb expects
something very defined inside and can't adapt easily to this, it's not
just an endianess problem, at least it will crash faster now avoid 10
error messages!