More dungeon texts
mainly wall texts, scrolls and characters titles.
This is done by searching texts in the dungeon.dat using csbuild the
copy text. If you have a french version of dm you can do the same in fr.
The st version 1.3 is a very poor translation by the way, all sorts of
mistakes inside... !