Finally allows csb to start without any dungeon.dat or graphics.dat
It's required for linux when the program is installed in a system
directory, then it will have to select the dungeon.dat from menu, it's
now possible.
For now you see the unexpected reproduction of an st crashing during
reset as the game screen under the menu while there is no data file
yet... I'll fix it another time!