change dungeon.dat selection logic
tested with conflux dungeon, and I noticed I forgot csbgraphics.dat in
the open hack for this function. This makes too many files for a hack,
so I removed the hack completely. Instead when selecting a dungeon.dat
it calls chdir in its directory so any open file will be opened here.
Then it just checks there is a graphics.dat. If not it calls another
file selector for this one. The 2 filenames are stored in char[]
variables to avoid to mess with allocations here.