the slow progress to accept the pc graphics.dat
it's purely experimental, it requires the 558 to 561.dat files in the
local directory to be able to initialize the datas correctly since
they're not in the pc version anymore.
Since the error about monters movement ticks = 0 is gone, it seems to
work.
Now it crashes on the expansion of the 1st graphic, the logo, and it's
probably going to be harder to fix, maybe next time... !